Rename and extract out icon test
